Update test and test name
